requirements.txt
streamlit==1.37.1
python-dotenv==1.0.1
langchain==0.2.14
langchain-community==0.2.12
langchain-openai==0.1.22
langchain-chroma==0.1.4
tiktoken==0.7.0
beautifulsoup4==4.12.2
openai>=1.40.0,<2.0.0
rank-bm25==0.2.2
